Corrected error

+++ LibraryProgram/src/main/java/com/swe/libraryprogram/service/LibraryAdminService.java
@@ -6,6 +6,7 @@ import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 
 public class LibraryAdminService extends UserService {
@@ -105,7 +106,11 @@ public class LibraryAdminService extends UserService {
 
     public Boolean removeElement(Element element) {
         try {
-            MainService.getInstance().getElementDAO().getElement(element.getId());
+            Element elewment = MainService.getInstance().getElementDAO().getElement(element.getId());
+            if(!Objects.equals(element.getQuantityAvailable(), element.getQuantity())){
+                System.out.println("Impossibile rimuovere un elemento ancora in prestito");
+                return false;
+            }
             MainService.getInstance().getElementDAO().removeElement(element.getId());
             return true;
         } catch (SQLException e) {
